split up project views

add coldfront core setting for PROJECT_UPDATE_FIELDS

UI Fixes
